+++ src/main/java/de/mkammerer/argon2/Argon2Constants.java
@@ -1,17 +1,9 @@
 package de.mkammerer.argon2;
 
-import de.mkammerer.argon2.jna.Argon2_type;
-
 /**
  * Commonly used constants.
  */
 public final class Argon2Constants {
-    /**
-     * Static class, no instances allowed.
-     */
-    private Argon2Constants() {
-    }
-
     /**
      * Salt length in bytes.
      */
@@ -21,4 +13,10 @@ public final class Argon2Constants {
      * Hash length in bytes.
      */
     public static final int DEFAULT_HASH_LENGTH = 32;
+
+    /**
+     * Static class, no instances allowed.
+     */
+    private Argon2Constants() {
+    }
 }