+++ src/test/java/de/mkammerer/argon2/test/LibraryTest.java
@@ -1,16 +1,20 @@
 package de.mkammerer.argon2.test;
 
 import de.mkammerer.argon2.Argon2;
+import de.mkammerer.argon2.Argon2Advanced;
 import de.mkammerer.argon2.Argon2Factory;
+import de.mkammerer.argon2.Argon2Factory.Argon2Types;
 import org.junit.Test;
 
 import java.nio.charset.Charset;
+import java.util.Random;
 
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.*;
 import static org.junit.Assert.assertThat;
 
 public class LibraryTest {
+    private final Random random = new Random();
+
     @Test
     public void testArgon2i() throws Exception {
         Argon2 sut = Argon2Factory.create(Argon2Factory.Argon2Types.ARGON2i);
@@ -154,5 +158,41 @@ public class LibraryTest {
         for (char c : array) {
             assertThat(c, is((char) 0));
         }
+    }
+    
+    @Test
+    public void testArgon2dRawHash() throws Exception {
+        byte[] salt = new byte[16];
+        random.nextBytes(salt);
+
+        Argon2Advanced sut = Argon2Factory.createAdvanced(Argon2Types.ARGON2d, salt.length, 32);
+        byte[] hash = sut.rawHash(2, 65536, 1, "password", salt);
+        
+        assertThat(sut.rawHash(2, 65536, 1, "password", salt), is(hash));
+        assertThat(sut.rawHash(2, 65536, 1, "not-the-password", salt), is(not(hash)));
+    }
+
+    @Test
+    public void testArgon2iRawHash() throws Exception {
+        byte[] salt = new byte[16];
+        random.nextBytes(salt);
+
+        Argon2Advanced sut = Argon2Factory.createAdvanced(Argon2Types.ARGON2i, salt.length, 32);
+        byte[] hash = sut.rawHash(2, 65536, 1, "password", salt);
+        
+        assertThat(sut.rawHash(2, 65536, 1, "password", salt), is(hash));
+        assertThat(sut.rawHash(2, 65536, 1, "not-the-password", salt), is(not(hash)));
+    }
+
+    @Test
+    public void testArgon2idRawHash() throws Exception {
+        byte[] salt = new byte[16];
+        random.nextBytes(salt);
+
+        Argon2Advanced sut = Argon2Factory.createAdvanced(Argon2Types.ARGON2id, salt.length, 32);
+        byte[] hash = sut.rawHash(2, 65536, 1, "password", salt);
+        
+        assertThat(sut.rawHash(2, 65536, 1, "password", salt), is(hash));
+        assertThat(sut.rawHash(2, 65536, 1, "not-the-password", salt), is(not(hash)));
     }
 }